Moved MagnitudeRange class to utils module
